requirements.txt
paramiko==3.4.0
click==8.1.7
python-dotenv==1.0.0 